Changing the whole RTree API to use float instead of double.
git-svn-id: http://chianti.ucsd.edu/svn@2144 0ecc0d97-ab19-0410-9704-bfe1a75892f5

+++ csplugins/trunk/ucsd/nlandys/rtree/src/cytoscape/geom/rtree/test/BasicQuietRTreeTest.java
@@ -1,3 +1,4 @@
+
 package cytoscape.geom.rtree.test;
 
 import cytoscape.geom.rtree.RTree;
@@ -13,17 +14,17 @@ public class BasicQuietRTreeTest
 
     for (int a = 0;; a++)
     { // BEGIN EMPTY TREE TESTS: We run our first tests when this tree empty.
-      double[] extentsArr = new double[4];
+      float[] extentsArr = new float[4];
       IntEnumerator iter = tree.queryOverlap
-        (Double.NEGATIVE_INFINITY, Double.NEGATIVE_INFINITY,
-         Double.POSITIVE_INFINITY, Double.POSITIVE_INFINITY, extentsArr, 0);
+        (Float.NEGATIVE_INFINITY, Float.NEGATIVE_INFINITY,
+         Float.POSITIVE_INFINITY, Float.POSITIVE_INFINITY, extentsArr, 0);
       if (iter.numRemaining() != 0)
         throw new IllegalStateException
           ("did not expect query to generate results");
-      if (extentsArr[0] != Double.POSITIVE_INFINITY ||
-          extentsArr[1] != Double.POSITIVE_INFINITY ||
-          extentsArr[2] != Double.NEGATIVE_INFINITY ||
-          extentsArr[3] != Double.NEGATIVE_INFINITY)
+      if (extentsArr[0] != Float.POSITIVE_INFINITY ||
+          extentsArr[1] != Float.POSITIVE_INFINITY ||
+          extentsArr[2] != Float.NEGATIVE_INFINITY ||
+          extentsArr[3] != Float.NEGATIVE_INFINITY)
         throw new IllegalStateException
           ("expected query to return inverted infinite extents");
 
@@ -41,13 +42,13 @@ public class BasicQuietRTreeTest
         for (int k = j * 1000; k < stop; k++) tree.delete(k); }
     } // END EMPTY TREE TESTS.
 
-    tree.insert(0, 0.0, 0.0, 1.0, 1.0);
-    tree.insert(1, 2.0, 2.0, 3.0, 3.0);
-    tree.insert(2, 0.5, 1.0, 1.5, 2.0);
+    tree.insert(0, 0.0f, 0.0f, 1.0f, 1.0f);
+    tree.insert(1, 2.0f, 2.0f, 3.0f, 3.0f);
+    tree.insert(2, 0.5f, 1.0f, 1.5f, 2.0f);
 
     for (int a = 0;; a++)
     { // BEGIN ROOT LEAF TEST: Still before any split.
-      double[] extentsArr = new double[5];
+      float[] extentsArr = new float[5];
       for (int i = 0; i < 3; i++)
         if (!tree.exists(i, extentsArr, 0))
           throw new IllegalStateException("entry " + i + " does not exist");
@@ -61,8 +62,8 @@ public class BasicQuietRTreeTest
         throw new IllegalStateException("tree's size() is not 3");
 
       IntEnumerator iter = tree.queryOverlap
-        (Double.NEGATIVE_INFINITY, Double.NEGATIVE_INFINITY,
-         Double.POSITIVE_INFINITY, Double.POSITIVE_INFINITY, extentsArr, 0);
+        (Float.NEGATIVE_INFINITY, Float.NEGATIVE_INFINITY,
+         Float.POSITIVE_INFINITY, Float.POSITIVE_INFINITY, extentsArr, 0);
       if (iter.numRemaining() != 3)
         throw new IllegalStateException("expected query to generate 3 hits");
       IntBTree cache = new IntBTree();
@@ -77,7 +78,7 @@ public class BasicQuietRTreeTest
           extentsArr[2] != 3.0 || extentsArr[3] != 3.0)
         throw new IllegalStateException("extents from query wrong");
 
-      iter = tree.queryOverlap(1.25, 2.0, 2.1, 3.3, extentsArr, 1);
+      iter = tree.queryOverlap(1.25f, 2.0f, 2.1f, 3.3f, extentsArr, 1);
       if (iter.numRemaining() != 2)
         throw new IllegalStateException("exptected query to return 2 hits");
       cache.insert(1); cache.insert(2); foo = 0;
@@ -98,11 +99,11 @@ public class BasicQuietRTreeTest
         for (int k = (j * 1000) + 3; k < stop; k++) tree.delete(k); }
     } // END ROOT LEAF TEST.
 
-    tree.insert(3, 2.5, 0.5, 3.5, 1.5);
+    tree.insert(3, 2.5f, 0.5f, 3.5f, 1.5f);
 
     for (int a = 0;; a++)
     { // BEGIN SIMPLE ROOT SPLIT TEST: Minimum # entries with a split.
-      double[] extentsArr = new double[4];
+      float[] extentsArr = new float[4];
       for (int i = 0; i < 4; i++)
         if (!tree.exists(i, extentsArr, 0))
           throw new IllegalStateException("entry " + i + " does not exist");
@@ -116,8 +117,8 @@ public class BasicQuietRTreeTest
         throw new IllegalStateException("tree's size() is not 4");
 
       IntEnumerator iter = tree.queryOverlap
-        (Double.NEGATIVE_INFINITY, Double.NEGATIVE_INFINITY,
-         Double.POSITIVE_INFINITY, Double.POSITIVE_INFINITY, extentsArr, 0);
+        (Float.NEGATIVE_INFINITY, Float.NEGATIVE_INFINITY,
+         Float.POSITIVE_INFINITY, Float.POSITIVE_INFINITY, extentsArr, 0);
       if (iter.numRemaining() != 4)
         throw new IllegalStateException("expected query to generate 4 hits");
       IntBTree cache = new IntBTree();
@@ -132,18 +133,18 @@ public class BasicQuietRTreeTest
           extentsArr[2] != 3.5 || extentsArr[3] != 3.0)
         throw new IllegalStateException("extents from query wrong");
 
-      iter = tree.queryOverlap(2.0, 0.5, 2.2, 1.9, extentsArr, 0);
+      iter = tree.queryOverlap(2.0f, 0.5f, 2.2f, 1.9f, extentsArr, 0);
       if (iter.numRemaining() != 0)
         throw new IllegalStateException("expected query to generate 0 hits");
-      if (extentsArr[0] != Double.POSITIVE_INFINITY ||
-          extentsArr[1] != Double.POSITIVE_INFINITY ||
-          extentsArr[2] != Double.NEGATIVE_INFINITY ||
-          extentsArr[3] != Double.NEGATIVE_INFINITY)
+      if (extentsArr[0] != Float.POSITIVE_INFINITY ||
+          extentsArr[1] != Float.POSITIVE_INFINITY ||
+          extentsArr[2] != Float.NEGATIVE_INFINITY ||
+          extentsArr[3] != Float.NEGATIVE_INFINITY)
         throw new IllegalStateException
           ("query extents - expected inverted infinite");
 
       iter = tree.queryOverlap
-        (Double.NEGATIVE_INFINITY, 1.1, Double.POSITIVE_INFINITY, 1.2,
+        (Float.NEGATIVE_INFINITY, 1.1f, Float.POSITIVE_INFINITY, 1.2f,
          extentsArr, 0);
       if (iter.numRemaining() != 2)
         throw new IllegalStateException("expected query to generate 2 hits");
@@ -157,7 +158,7 @@ public class BasicQuietRTreeTest
           extentsArr[2] != 3.5 || extentsArr[3] != 2.0)
         throw new IllegalStateException("extents from query wrong");
 
-      iter = tree.queryOverlap(1.0, 1.0, 1.0, 1.0, extentsArr, 0);
+      iter = tree.queryOverlap(1.0f, 1.0f, 1.0f, 1.0f, extentsArr, 0);
       if (iter.numRemaining() != 2)
         throw new IllegalStateException("expected query to generate 2 hits");
       cache.insert(0); cache.insert(2); foo = 0;
@@ -180,34 +181,34 @@ public class BasicQuietRTreeTest
 
     { // BEGIN EXCEPTION HANDLING TEST.
       boolean exceptionCaught = false;
-      try { tree.insert(0, 0.0, 0.0, 1.0, 1.0); }
+      try { tree.insert(0, 0.0f, 0.0f, 1.0f, 1.0f); }
       catch (IllegalStateException e) { exceptionCaught = true; }
       if (!exceptionCaught) throw new IllegalStateException
                               ("expected exception for duplicate objKey");
 
       exceptionCaught = false;
-      try { tree.insert(-1, 0.0, 0.0, 1.0, 1.0); }
+      try { tree.insert(-1, 0.0f, 0.0f, 1.0f, 1.0f); }
       catch (IllegalArgumentException e) { exceptionCaught = true; }
       if (!exceptionCaught) throw new IllegalStateException
                               ("expected exception for negative objKey");
 
       exceptionCaught = false;
-      try { tree.insert(5, 1.0, 1.0, 0.0, 0.0); }
+      try { tree.insert(5, 1.0f, 1.0f, 0.0f, 0.0f); }
       catch (IllegalArgumentException e) { exceptionCaught = true; }
       if (!exceptionCaught) throw new IllegalStateException
                               ("expected exception for min > max");
     } // END EXCEPTION HANDLING TEST.
 
-    tree.insert(4, 3.0, -0.25, 4.0, 0.75);
-    tree.insert(5, -0.5, 2.5, 0.5, 3.5);
-    tree.insert(6, 2.75, 2.25, 3.75, 3.25);
-    tree.insert(7, 1.25, 1.75, 2.25, 2.75);
-    tree.insert(8, 1.0, 6.0, 2.0, 7.0);
-    tree.insert(9, -2.0, 1.0, -1.0, 2.0);
+    tree.insert(4, 3.0f, -0.25f, 4.0f, 0.75f);
+    tree.insert(5, -0.5f, 2.5f, 0.5f, 3.5f);
+    tree.insert(6, 2.75f, 2.25f, 3.75f, 3.25f);
+    tree.insert(7, 1.25f, 1.75f, 2.25f, 2.75f);
+    tree.insert(8, 1.0f, 6.0f, 2.0f, 7.0f);
+    tree.insert(9, -2.0f, 1.0f, -1.0f, 2.0f);
 
     for (int a = 0;; a++)
     { // BEGIN DEPTH THREE TEST.
-      double[] extentsArr = new double[4];
+      float[] extentsArr = new float[4];
       for (int i = 9; i >= 0; i--)
         if (!tree.exists(i, extentsArr, 0))
           throw new IllegalStateException("entry " + i + " does not exist");
@@ -221,8 +222,8 @@ public class BasicQuietRTreeTest
         throw new IllegalStateException("tree's size() is not 10");
 
       IntEnumerator iter = tree.queryOverlap
-        (Double.NEGATIVE_INFINITY, Double.NEGATIVE_INFINITY,
-         Double.POSITIVE_INFINITY, Double.POSITIVE_INFINITY, extentsArr, 0);
+        (Float.NEGATIVE_INFINITY, Float.NEGATIVE_INFINITY,
+         Float.POSITIVE_INFINITY, Float.POSITIVE_INFINITY, extentsArr, 0);
       if (iter.numRemaining() != 10)
         throw new IllegalStateException("expected query to generate 10 hits");
       IntBTree cache = new IntBTree();
@@ -237,7 +238,7 @@ public class BasicQuietRTreeTest
           extentsArr[2] != 4.0 || extentsArr[3] != 7.0)
         throw new IllegalStateException("extents from query wrong");
 
-      iter = tree.queryOverlap(1.0, 1.25, 3.0, 5.0, extentsArr, 0);
+      iter = tree.queryOverlap(1.0f, 1.25f, 3.0f, 5.0f, extentsArr, 0);
       if (iter.numRemaining() != 5)
         throw new IllegalStateException("expected query to generate 5 hits");
       cache.insert(1); cache.insert(2); cache.insert(3); cache.insert(6);
@@ -251,7 +252,7 @@ public class BasicQuietRTreeTest
           extentsArr[2] != 3.75 || extentsArr[3] != 3.25)
         throw new IllegalStateException("extents from query wrong");
 
-      iter = tree.queryOverlap(-1.5, 0.25, 0.25, 3.0, extentsArr, 0);
+      iter = tree.queryOverlap(-1.5f, 0.25f, 0.25f, 3.0f, extentsArr, 0);
       if (iter.numRemaining() != 3)
         throw new IllegalStateException("expected query to generate 3 hits");
       cache.insert(0); cache.insert(5); cache.insert(9); foo = 0;
@@ -264,7 +265,7 @@ public class BasicQuietRTreeTest
           extentsArr[2] != 1.0 || extentsArr[3] != 3.5)
         throw new IllegalStateException("extents from query wrong");
 
-      iter = tree.queryOverlap(1.5, 6.5, 1.5, 6.5, extentsArr, 0);
+      iter = tree.queryOverlap(1.5f, 6.5f, 1.5f, 6.5f, extentsArr, 0);
       if (iter.numRemaining() != 1)
         throw new IllegalStateException("expected query to generate 1 hit");
       cache.insert(8); foo = 0;
@@ -277,15 +278,16 @@ public class BasicQuietRTreeTest
           extentsArr[2] != 2.0 || extentsArr[3] != 7.0)
         throw new IllegalStateException("extents from query wrong");
 
-      iter = tree.queryOverlap(3.0, 5.0, 8.0, 9.0, extentsArr, 0);
+      iter = tree.queryOverlap(3.0f, 5.0f, 8.0f, 9.0f, extentsArr, 0);
       if (iter.numRemaining() != 0)
         throw new IllegalStateException("did not expect any query hits");
 
-      iter = tree.queryOverlap(-100.0, -100.0, -99.0, -99.0, extentsArr, 0);
+      iter = tree.queryOverlap(-100.0f, -100.0f, -99.0f, -99.0f,
+                               extentsArr, 0);
       if (iter.numRemaining() != 0)
         throw new IllegalStateException("did not expect any query hits");
 
-      iter = tree.queryOverlap(-1.0, 0.75, 3.0, 6.0, extentsArr, 0);
+      iter = tree.queryOverlap(-1.0f, 0.75f, 3.0f, 6.0f, extentsArr, 0);
       if (iter.numRemaining() != 10)
         throw new IllegalStateException("expected 10 query hits");
 
@@ -298,24 +300,24 @@ public class BasicQuietRTreeTest
     } // END DEPTH THREE TEST.
 
     for (int a = 0;; a++) {
-      tree.insert(10, 2.0, 4.0, 3.0, 5.0);
-      tree.insert(11, 1.5, 3.75, 3.5, 4.25);
-      tree.insert(12, 2.5, 3.5, 3.0, 5.5);
-      tree.insert(13, -4.0, 6.0, -2.0, 8.0);
-      tree.insert(14, -4.25, 5.75, 2.25, 8.25);
-      tree.insert(15, 2.0, -1.0, 2.0, -1.0);
-      tree.insert(16, -1.25, 0.5, -1.25, 3.0);
-      tree.insert(17, -0.5, -0.5, 1.5, 0.5);
-      tree.insert(18, 0.25, 4.0, 1.25, 5.0);
-      tree.insert(19, 4.0, 1.0, 5.0, 2.0);
-      tree.insert(20, 4.0, 3.0, 5.0, 4.0);
-      tree.insert(21, 4.25, -1.5, 4.75, 5.0);
-      tree.insert(22, 3.0, -1.75, 5.0, -1.0);
-      tree.insert(23, 1.25, 0.25, 2.25, 1.25);
-      tree.insert(24, -2.0, 9.0, -1.0, 10.0);
-      tree.insert(25, 1.0, 9.0, 2.0, 10.0);
-      tree.insert(26, -2.0, 5.0, -1.0, 6.0);
-      tree.insert(27, -2.5, 5.25, -1.75, 9.25);
+      tree.insert(10, 2.0f, 4.0f, 3.0f, 5.0f);
+      tree.insert(11, 1.5f, 3.75f, 3.5f, 4.25f);
+      tree.insert(12, 2.5f, 3.5f, 3.0f, 5.5f);
+      tree.insert(13, -4.0f, 6.0f, -2.0f, 8.0f);
+      tree.insert(14, -4.25f, 5.75f, 2.25f, 8.25f);
+      tree.insert(15, 2.0f, -1.0f, 2.0f, -1.0f);
+      tree.insert(16, -1.25f, 0.5f, -1.25f, 3.0f);
+      tree.insert(17, -0.5f, -0.5f, 1.5f, 0.5f);
+      tree.insert(18, 0.25f, 4.0f, 1.25f, 5.0f);
+      tree.insert(19, 4.0f, 1.0f, 5.0f, 2.0f);
+      tree.insert(20, 4.0f, 3.0f, 5.0f, 4.0f);
+      tree.insert(21, 4.25f, -1.5f, 4.75f, 5.0f);
+      tree.insert(22, 3.0f, -1.75f, 5.0f, -1.0f);
+      tree.insert(23, 1.25f, 0.25f, 2.25f, 1.25f);
+      tree.insert(24, -2.0f, 9.0f, -1.0f, 10.0f);
+      tree.insert(25, 1.0f, 9.0f, 2.0f, 10.0f);
+      tree.insert(26, -2.0f, 5.0f, -1.0f, 6.0f);
+      tree.insert(27, -2.5f, 5.25f, -1.75f, 9.25f);
 
       if (a == 1) break;
       for (int k = 10; k < 28; k++) tree.delete(k); }
@@ -323,7 +325,7 @@ public class BasicQuietRTreeTest
 
     for (int a = 0;; a++)
     { // BEGIN DEPTH FOUR TEST.
-      double[] extentsArr = new double[4];
+      float[] extentsArr = new float[4];
       for (int i = 27; i >= 0; i--)
         if (!tree.exists(i, extentsArr, 0))
           throw new IllegalStateException("entry " + i + " does not exist");
@@ -339,8 +341,8 @@ public class BasicQuietRTreeTest
         throw new IllegalStateException("tree's size() is not 28");
 
       IntEnumerator iter = tree.queryOverlap
-        (Double.NEGATIVE_INFINITY, Double.NEGATIVE_INFINITY,
-         Double.POSITIVE_INFINITY, Double.POSITIVE_INFINITY, extentsArr, 0);
+        (Float.NEGATIVE_INFINITY, Float.NEGATIVE_INFINITY,
+         Float.POSITIVE_INFINITY, Float.POSITIVE_INFINITY, extentsArr, 0);
       if (iter.numRemaining() != 28)
         throw new IllegalStateException("expected query to give 28 hits");
       IntBTree cache = new IntBTree();
@@ -355,7 +357,7 @@ public class BasicQuietRTreeTest
           extentsArr[2] != 5.0 || extentsArr[3] != 10.0)
         throw new IllegalStateException("extents from query wrong");
 
-      iter = tree.queryOverlap(-2.0, 6.0, -2.0, 6.0, extentsArr, 0);
+      iter = tree.queryOverlap(-2.0f, 6.0f, -2.0f, 6.0f, extentsArr, 0);
       if (iter.numRemaining() != 4)
         throw new IllegalStateException("expected query to generate 4 hits");
       cache.insert(13); cache.insert(14); cache.insert(26); cache.insert(27);
@@ -369,7 +371,7 @@ public class BasicQuietRTreeTest
           extentsArr[2] != 2.25 || extentsArr[3] != 9.25)
         throw new IllegalStateException("extents from query wrong");
 
-      iter = tree.queryOverlap(2.5, 3.75, 6.0, 6.0, extentsArr, 0);
+      iter = tree.queryOverlap(2.5f, 3.75f, 6.0f, 6.0f, extentsArr, 0);
       if (iter.numRemaining() != 5)
         throw new IllegalStateException("expected query to generate 5 hits");
       cache.insert(10); cache.insert(11); cache.insert(12); cache.insert(20);
@@ -383,7 +385,7 @@ public class BasicQuietRTreeTest
           extentsArr[2] != 5.0 || extentsArr[3] != 5.5)
         throw new IllegalStateException("extents from query wrong");
 
-      iter = tree.queryOverlap(1.75, -1.5, 3.25, -0.5, extentsArr, 0);
+      iter = tree.queryOverlap(1.75f, -1.5f, 3.25f, -0.5f, extentsArr, 0);
       if (iter.numRemaining() != 2)
         throw new IllegalStateException("expected query to generate 2 hits");
       cache.insert(15); cache.insert(22); foo = 0;
@@ -396,7 +398,7 @@ public class BasicQuietRTreeTest
           extentsArr[2] != 5.0 || extentsArr[3] != -1.0)
         throw new IllegalStateException("extents from query wrong");
 
-      iter = tree.queryOverlap(-3.0, -5.0, 7.0, 2.75, extentsArr, 0);
+      iter = tree.queryOverlap(-3.0f, -5.0f, 7.0f, 2.75f, extentsArr, 0);
       if (iter.numRemaining() != 16)
         throw new IllegalStateException("expected query to generate 16 hits");
       cache.insert(0); cache.insert(1); cache.insert(2); cache.insert(3);
@@ -413,7 +415,7 @@ public class BasicQuietRTreeTest
           extentsArr[2] != 5.0 || extentsArr[3] != 5.0)
         throw new IllegalStateException("extents from query wrong");
 
-      iter = tree.queryOverlap(1.0, 4.5, 3.0, 6.25, extentsArr, 0);
+      iter = tree.queryOverlap(1.0f, 4.5f, 3.0f, 6.25f, extentsArr, 0);
       if (iter.numRemaining() != 5)
         throw new IllegalStateException("expected query to generate 5 hits");
       cache.insert(8); cache.insert(10); cache.insert(12); cache.insert(14);
@@ -427,7 +429,7 @@ public class BasicQuietRTreeTest
           extentsArr[2] != 3.0 || extentsArr[3] != 8.25)
         throw new IllegalStateException("extents from query wrong");
 
-      iter = tree.queryOverlap(0.75, 0.25, 1.5, 0.75, extentsArr, 0);
+      iter = tree.queryOverlap(0.75f, 0.25f, 1.5f, 0.75f, extentsArr, 0);
       if (iter.numRemaining() != 3)
         throw new IllegalStateException("expected query to generate 3 hits");
       cache.insert(0); cache.insert(17); cache.insert(23); foo = 0;
@@ -440,7 +442,7 @@ public class BasicQuietRTreeTest
           extentsArr[2] != 2.25 || extentsArr[3] != 1.25)
         throw new IllegalStateException("extents from query wrong");
 
-      iter = tree.queryOverlap(1.75, -10.0, 1.75, 30.0, extentsArr, 0);
+      iter = tree.queryOverlap(1.75f, -10.0f, 1.75f, 30.0f, extentsArr, 0);
       if (iter.numRemaining() != 6)
         throw new IllegalStateException("expected query to generate 6 hits");
       cache.insert(7); cache.insert(8); cache.insert(11); cache.insert(14);
@@ -454,15 +456,15 @@ public class BasicQuietRTreeTest
           extentsArr[2] != 3.5 || extentsArr[3] != 10.0)
         throw new IllegalStateException("extents from query wrong");
 
-      iter = tree.queryOverlap(0.75, 3.0, 1.75, 3.5, extentsArr, 0);
+      iter = tree.queryOverlap(0.75f, 3.0f, 1.75f, 3.5f, extentsArr, 0);
       if (iter.numRemaining() != 0)
         throw new IllegalStateException("did not expect query results");
 
-      iter = tree.queryOverlap(2.5, 1.75, 3.75, 1.75, null, -1);
+      iter = tree.queryOverlap(2.5f, 1.75f, 3.75f, 1.75f, null, -1);
       if (iter.numRemaining() != 0)
         throw new IllegalStateException("did not expect query results");
 
-      iter = tree.queryOverlap(-2.0, -1.0, 4.25, 9.0, extentsArr, 0);
+      iter = tree.queryOverlap(-2.0f, -1.0f, 4.25f, 9.0f, extentsArr, 0);
       if (iter.numRemaining() != 28)
         throw new IllegalStateException("expected 28 (all) query hits");
       for (int i = 0; i < 28; i++) cache.insert(i);
